feat: implemented modules solution
